\name{countsNigerian}

\alias{countsNigerian}
\alias{genderNigerian}

\docType{data}
\title{
RNA-seq count data
}
\description{
RNA-seq count data for Nigerian populations. In the case of Nigerian, gender information is available in the 'genderNigerian' vector.
}
\usage{
data(countsNigerian)
data(genderNigerian)
}
\format{
  
  'countsNigerian': data.frame containing 38415 gene counts (rows) for 69 Nigerian
  individuals.

  'genderNigerian': factor vector providing gender information about the Nigerian individuals.

}

\details{
Count data information was obtained from datasets gathered by
Pickrell et al. (2010).  These data were obtained from mapping
the reads to the human reference genome and summarizing the counts
using Ensemble release 63. Readcounts were generated by selecting those reads that
mapped to annotated genes, either completely within an exon or spanning
a junction. The annotation used was GENCODE version 3c.
}

\source{
J.K. Pickrell, J.C. Marioni, A.A Pai, J.F. Degner, B.E. Engelhardt,
E. Nkadori, J.B Veyrieras, M. Stephens, Y. Gilad, and J.K. Pritchard
Nature 464:768-772, 2010.
}
\references{
  M. Esnaola, P. Puig, D. Gonzalez, R. Castelo, J.R. Gonzalez.
  Gene-specific count data distributions are required in RNA-seq
  experiments with extensive replication. Submitted.
}
\seealso{
  \code{\link{countsNigerianNorm}}
  \code{\link{annotNigerian}}
  }
\examples{
data(countsNigerian)
}
\keyword{datasets}
